Adding a new category to the collections menu
The thing I wanted to do in settings isn't going to work the way I want. I'm instead suggesting to move the checks to the collection database under a new category. I don't know if this will break anything for either builder, so I'd like you to have a look at it before pushing anything